added acpi error fix for few laptops
The introduction of Optimus laptops have created new challenges for Linux users. One of the biggest issues is having the proprietary graphics card work when the system boots. Many times the advice given is to add an Advanced Configuration and Power Interface (acpi ) kernel parameter to grub. While a specific acpi kernel parameter may be given, it is sometimes not appropriate to the hardware it is being applied to. This can cause system instability and some functions of the laptop to not work properly. Once the user does find a parameter that works, they are told to add it to their grub config file so it will be automatically load at boot.
The kernel parameter apci_osi=Linux can be used in newer (late 2019 and later) laptops to fix an issue where your laptop just boots to a black screen and the fans start to run constantly, An error message about configuring “Backlight”, the external HDMI port is not working.